OGM-366 Add remote Neo4j integration

+++ neo4j/src/test/java/org/hibernate/ogm/datastore/neo4j/utils/Neo4jTestHelper.java
@@ -9,6 +9,7 @@ package org.hibernate.ogm.datastore.neo4j.utils;
 import java.io.File;
 import java.io.IOException;
 import java.io.InputStream;
+import java.util.Arrays;
 import java.util.Collections;
 import java.util.HashMap;
 import java.util.List;
@@ -25,9 +26,11 @@ import org.hibernate.ogm.datastore.document.options.AssociationStorageType;
 import org.hibernate.ogm.datastore.neo4j.Neo4j;
 import org.hibernate.ogm.datastore.neo4j.Neo4jDialect;
 import org.hibernate.ogm.datastore.neo4j.Neo4jProperties;
+import org.hibernate.ogm.datastore.neo4j.RemoteNeo4jDialect;
 import org.hibernate.ogm.datastore.neo4j.dialect.impl.NodeLabel;
 import org.hibernate.ogm.datastore.neo4j.impl.Neo4jDatastoreProvider;
 import org.hibernate.ogm.datastore.neo4j.remote.impl.Neo4jClient;
+import org.hibernate.ogm.datastore.neo4j.remote.impl.RemoteNeo4jDatastoreProvider;
 import org.hibernate.ogm.datastore.neo4j.remote.json.impl.ErrorResponse;
 import org.hibernate.ogm.datastore.neo4j.remote.json.impl.Statement;
 import org.hibernate.ogm.datastore.neo4j.remote.json.impl.Statements;
@@ -56,7 +59,7 @@ public class Neo4jTestHelper implements TestableGridDialect {
 
 	private static final Map<String, String> hibernateProperties = readProperties();
 
-	private static final String ROOT_FOLDER = hibernateProperties.get( Neo4jProperties.DATABASE_PATH ) + File.separator + "NEO4J";
+	private static final String ROOT_FOLDER = hibernateProperties.get( Neo4jProperties.DATABASE_PATH ) + File.separator + "NEO4J_REMOTE";
 
 	/**
 	 * Query for counting all entities. This takes embedded entities and temporary nodes (which never should show up
@@ -103,11 +106,19 @@ public class Neo4jTestHelper implements TestableGridDialect {
 	}
 
 	private long getNumberOfEntities(Session session, DatastoreProvider provider) {
-		GraphDatabaseService graphDb = ( (Neo4jDatastoreProvider) provider ).getDatabase();
-		ResourceIterator<Long> result = graphDb.execute( ENTITY_COUNT_QUERY ).columnAs( "count" );
-		Long count = result.next();
-		result.close();
-		return count.longValue();
+		if ( isRemote( provider ) ) {
+			Neo4jClient remoteNeo4j = ( (RemoteNeo4jDatastoreProvider) provider ).getDatabase();
+			Statement statement = new Statement( ENTITY_COUNT_QUERY );
+			statement.setResultDataContents( Arrays.asList( Statement.AS_ROW ) );
+			return readCountFromResponse( session, remoteNeo4j, statement );
+		}
+		else {
+			GraphDatabaseService graphDb = ( (Neo4jDatastoreProvider) provider ).getDatabase();
+			ResourceIterator<Long> result = graphDb.execute( ENTITY_COUNT_QUERY ).columnAs( "count" );
+			Long count = result.next();
+			result.close();
+			return count.longValue();
+		}
 	}
 
 	@Override
@@ -123,11 +134,19 @@ public class Neo4jTestHelper implements TestableGridDialect {
 	}
 
 	private long getNumberOfAssociations(Session session, BaseDatastoreProvider provider) {
-		GraphDatabaseService graphDb = ( (Neo4jDatastoreProvider) provider ).getDatabase();
-		ResourceIterator<Long> result = graphDb.execute( ASSOCIATION_COUNT_QUERY ).columnAs( "count" );
-		Long count = result.next();
-		result.close();
-		return count.longValue();
+		if ( isRemote( provider ) ) {
+			Neo4jClient remoteNeo4j = ( (RemoteNeo4jDatastoreProvider) provider ).getDatabase();
+			Statement statement = new Statement( ASSOCIATION_COUNT_QUERY );
+			statement.setResultDataContents( Arrays.asList( Statement.AS_ROW ) );
+			return readCountFromResponse( session, remoteNeo4j, statement );
+		}
+		else {
+			GraphDatabaseService graphDb = ( (Neo4jDatastoreProvider) provider ).getDatabase();
+			ResourceIterator<Long> result = graphDb.execute( ASSOCIATION_COUNT_QUERY ).columnAs( "count" );
+			Long count = result.next();
+			result.close();
+			return count.longValue();
+		}
 	}
 
 	private long readCountFromResponse(Session session, Neo4jClient remoteNeo4j, Statement statement) {
@@ -183,13 +202,31 @@ public class Neo4jTestHelper implements TestableGridDialect {
 	@Override
 	public void dropSchemaAndDatabase(SessionFactory sessionFactory) {
 		DatastoreProvider provider = getProvider( sessionFactory );
-		( (Stoppable) getProvider( sessionFactory ) ).stop();
-		Files.delete( new File( ROOT_FOLDER ) );
+		if ( provider instanceof RemoteNeo4jDatastoreProvider ) {
+			RemoteNeo4jDatastoreProvider remoteProvider = (RemoteNeo4jDatastoreProvider) provider;
+			Statements statements = new Statements();
+			statements.addStatement( DELETE_ALL );
+			remoteProvider.getDatabase().executeQueriesInNewTransaction( statements );
+		}
+		else {
+			( (Stoppable) getProvider( sessionFactory ) ).stop();
+			Files.delete( new File( ROOT_FOLDER ) );
+		}
 	}
 
 	public static void deleteAllElements(DatastoreProvider provider) {
-		GraphDatabaseService graphDb = ( (Neo4jDatastoreProvider) provider ).getDataBase();
-		graphDb.execute( DELETE_ALL ).close();
+		if ( isRemote( provider ) ) {
+			Neo4jClient remoteNeo4j = ( (RemoteNeo4jDatastoreProvider) provider ).getDatabase();
+			Statement statement = new Statement( DELETE_ALL );
+			statement.setResultDataContents( Arrays.asList( Statement.AS_ROW ) );
+			Statements statements = new Statements();
+			statements.addStatement( statement );
+			remoteNeo4j.executeQueriesInNewTransaction( statements );
+		}
+		else {
+			GraphDatabaseService graphDb = ( (Neo4jDatastoreProvider) provider ).getDatabase();
+			graphDb.execute( DELETE_ALL ).close();
+		}
 	}
 
 	@Override
@@ -241,6 +278,9 @@ public class Neo4jTestHelper implements TestableGridDialect {
 		if ( Neo4jDatastoreProvider.class.isInstance( provider ) ) {
 			return Neo4jDatastoreProvider.class.cast( provider );
 		}
+		if ( RemoteNeo4jDatastoreProvider.class.isInstance( provider ) ) {
+			return RemoteNeo4jDatastoreProvider.class.cast( provider );
+		}
 		throw new RuntimeException( "Not testing with Neo4jDB, cannot extract underlying provider" );
 	}
 
@@ -264,7 +304,14 @@ public class Neo4jTestHelper implements TestableGridDialect {
 		if ( Neo4jDatastoreProvider.class.isInstance( datastoreProvider ) ) {
 			return new Neo4jDialect( (Neo4jDatastoreProvider) datastoreProvider );
 		}
+		if ( RemoteNeo4jDatastoreProvider.class.isInstance( datastoreProvider ) ) {
+			return new RemoteNeo4jDialect( (RemoteNeo4jDatastoreProvider) datastoreProvider );
+		}
 		throw new RuntimeException( "Not testing with Neo4jDB, cannot extract underlying dialect" );
+	}
+
+	private static boolean isRemote(DatastoreProvider provider) {
+		return provider instanceof RemoteNeo4jDatastoreProvider;
 	}
 
 	@JsonIgnoreProperties(ignoreUnknown = true)
